InContext v0.1.0 InContext.Context

Contextual Views on Graphs

@todo (longterm): Write map and reduce functions

Types

edge_triple()

edge_triple() :: {node_id(), node_id(), number()}

node_id()

node_id() :: term()

t() (opaque)

t()

Functions

edge_list(ctx)

edge_list(InContext.Context.t()) :: {[edge_triple()], [edge_triple()]}

Returns implementation-indpendent view of edges in a context.

view(graph, node)

view(InContext.Graph.t(), node_id() | nil) ::
  {InContext.Context.t() | nil, InContext.Graph.t()}

Creates a view from a graph at a given node, and returns the rest of the graph

Examples

iex> use InContext
iex> graph = Graph.new |>
...>   Graph.add_edge(1,2) |>
...>   Graph.add_edge(2,3) |>
...>   Graph.add_edge(1,3)
iex> {ctx, graph2} = Context.view(graph, 2)
iex> ctx.node
2
iex> Context.edge_list(ctx)
{ [{1, 2, 1.0}], [{2, 3, 1.0}] }
iex> Graph.has_node?(graph, 2)
:true
iex> Graph.has_node?(graph2, 2)
:false