InContext v0.1.0 InContext.Graph.Edge

Weighted, Directed Graph Edges

Functions

as_pair(edge)

Returns edge as a pair without weight, for implementation-independent representations.

as_triple(edge)

Returns edge as a triple, for implementation-independent representations.

from(edge)

Retrieve from node from given Edge.

new(from, to, weight \\ 1.0)

Returns new edge from from to to, with optional weight.

to(edge)

Retrieve to node from given Edge.

weight(edge)

Retrieve weight from given Edge.