InContext v0.1.0 InContext.Search

Searching Algorithms using Graph Contexts

Types

node_id()

node_id() :: term()

Functions

bfs(graph, from, to)

bfs(InContext.Graph.t(), node_id(), node_id()) :: [node_id()]

Breadth-first search of a graph, returning nodes visited.

Examples

For the following graph:

0 -> [1 -> [3 -> [7,
                  8],
            4 -> [9,
                  10]],
      2 -> [5,
            6]]

iex> use InContext
iex> graph = Graph.tree_graph(4)
iex> Search.bfs(graph, 0, 3)
[{nil, 0}, {0, 1}, {0, 2}, {1, 3}]
iex> Search.bfs(graph, 0, 2)
[{nil, 0}, {0, 1}, {0, 2}]
iex> Search.bfs(graph, 1, 2)
[]
iex> Search.bfs(graph, 0, 100)
[]

dfs(graph, from, to)

dfs(InContext.Graph.t(), node_id(), node_id()) :: [node_id()]

Depth-first search of a graph, returning pairs of nodes visited with the node used to reach them.

Examples

For the following graph:

0 -> [1 -> [3 -> [7,
                  8],
            4 -> [9,
                  10]],
      2 -> [5,
            6]]

iex> use InContext
iex> graph = Graph.tree_graph(4)
iex> Search.dfs(graph, 0, 3)
[{nil, 0}, {0, 1}, {1, 3}]
iex> Search.dfs(graph, 0, 2)
[{nil, 0}, {0, 1}, {1, 3}, {3, 7}, {3, 8}, {1, 4}, {4, 9}, {4, 10}, {0, 2}]
iex> Search.dfs(graph, 1, 2)
[]
iex> Search.dfs(graph, 0, 100)
[]